+++ assets-async/src/main/kotlin/ktx/assets/async/AssetManagerWrapper.kt
@@ -6,10 +6,10 @@ import com.badlogic.gdx.assets.AssetLoaderParameters
 import com.badlogic.gdx.assets.AssetManager
 import com.badlogic.gdx.assets.loaders.AssetLoader
 import com.badlogic.gdx.assets.loaders.FileHandleResolver
-import com.badlogic.gdx.utils.Array as GdxArray
 import com.badlogic.gdx.utils.Logger
 import kotlinx.coroutines.launch
 import ktx.async.KtxAsync
+import com.badlogic.gdx.utils.Array as GdxArray
 
 /**
  * Extends [AssetManager], delegating all of its asset-related method calls to [assetStorage].
@@ -35,8 +35,10 @@ internal class AssetManagerWrapper(
 
   override fun clear() = dispose()
 
-  @Deprecated("This operation is non-blocking. Assets might still be loaded after this call.",
-    replaceWith = ReplaceWith("AssetStorage.dispose"))
+  @Deprecated(
+    "This operation is non-blocking. Assets might still be loaded after this call.",
+    replaceWith = ReplaceWith("AssetStorage.dispose")
+  )
   override fun dispose() {
     if (initiated) {
       logger.error("Not fully supported AssetManagerWrapper.dispose called by AssetLoader.")
@@ -48,14 +50,18 @@ internal class AssetManagerWrapper(
     }
   }
 
-  @Deprecated("Not supported by AssetStorage.",
-    replaceWith = ReplaceWith("contains(fileName, type)"))
+  @Deprecated(
+    "Not supported by AssetStorage.",
+    replaceWith = ReplaceWith("contains(fileName, type)")
+  )
   override fun contains(fileName: String): Boolean = false
   override fun contains(fileName: String, type: Class<*>?): Boolean =
     assetStorage.contains(AssetDescriptor(fileName, type))
 
-  @Deprecated("This operation is non-blocking. Assets might not be available in storage after call.",
-    replaceWith = ReplaceWith("AssetStorage.add"))
+  @Deprecated(
+    "This operation is non-blocking. Assets might not be available in storage after call.",
+    replaceWith = ReplaceWith("AssetStorage.add")
+  )
   override fun <T : Any> addAsset(fileName: String, type: Class<T>, asset: T) {
     logger.error("Not fully supported AssetManagerWrapper.addAsset called by AssetLoader.")
     KtxAsync.launch {
@@ -93,9 +99,11 @@ internal class AssetManagerWrapper(
   @Deprecated("Not supported by AssetStorage.", replaceWith = ReplaceWith("Nothing"))
   override fun getAssetNames(): GdxArray<String> = throw UnsupportedMethodException("getAssetNames")
 
-  @Deprecated("Multiple assets with different types can be listed under the same path.",
-    replaceWith = ReplaceWith("Nothing"))
-  override fun getAssetType(fileName: String): Class<*>? = throw UnsupportedMethodException("getAssetType")
+  @Deprecated(
+    "Multiple assets with different types can be listed under the same path.",
+    replaceWith = ReplaceWith("Nothing")
+  )
+  override fun getAssetType(fileName: String): Class<*> = throw UnsupportedMethodException("getAssetType")
 
   @Deprecated("Not supported by AssetStorage.", replaceWith = ReplaceWith("Nothing"))
   override fun <Asset : Any> getAll(type: Class<Asset>, out: GdxArray<Asset>): GdxArray<Asset> =
@@ -121,14 +129,18 @@ internal class AssetManagerWrapper(
   override fun isLoaded(assetDesc: AssetDescriptor<*>): Boolean = assetStorage.isLoaded(assetDesc)
   override fun isLoaded(fileName: String, type: Class<*>): Boolean = isLoaded(AssetDescriptor(fileName, type))
 
-  @Deprecated("Not supported by AssetStorage.",
-    replaceWith = ReplaceWith("isLoaded(fileName, type)"))
+  @Deprecated(
+    "Not supported by AssetStorage.",
+    replaceWith = ReplaceWith("isLoaded(fileName, type)")
+  )
   override fun isLoaded(fileName: String): Boolean = false.also {
     logger.error("Not supported AssetManagerWrapper.addAsset called by AssetLoader.")
   }
 
-  @Deprecated("AssetStorage requires type of asset to unload.",
-    replaceWith = ReplaceWith("AssetStorage.unload"))
+  @Deprecated(
+    "AssetStorage requires type of asset to unload.",
+    replaceWith = ReplaceWith("AssetStorage.unload")
+  )
   override fun unload(fileName: String) {
     logger.error("Not supported AssetManagerWrapper.unload called by AssetLoader.")
   }
@@ -143,15 +155,19 @@ internal class AssetManagerWrapper(
     }
   }
 
-  @Deprecated("AssetLoader instances can be mutable." +
-    "AssetStorage requires functional providers of loaders rather than singular instances.",
-    replaceWith = ReplaceWith("AssetStorage.setLoader"))
+  @Deprecated(
+    "AssetLoader instances can be mutable." +
+      "AssetStorage requires functional providers of loaders rather than singular instances.",
+    replaceWith = ReplaceWith("AssetStorage.setLoader")
+  )
   override fun <T : Any?, P : AssetLoaderParameters<T>?> setLoader(type: Class<T>, loader: AssetLoader<T, P>) =
     setLoader(type, null, loader)
 
-  @Deprecated("AssetLoader instances can be mutable." +
-    "AssetStorage requires functional providers of loaders rather than singular instances.",
-    replaceWith = ReplaceWith("AssetStorage.setLoader"))
+  @Deprecated(
+    "AssetLoader instances can be mutable." +
+      "AssetStorage requires functional providers of loaders rather than singular instances.",
+    replaceWith = ReplaceWith("AssetStorage.setLoader")
+  )
   override fun <T : Any?, P : AssetLoaderParameters<T>?> setLoader(
     type: Class<T>,
     suffix: String?,
@@ -164,14 +180,18 @@ internal class AssetManagerWrapper(
     }
   }
 
-  @Deprecated("AssetStorage requires type to find dependencies.",
-    replaceWith = ReplaceWith("AssetStorage.getDependencies"))
+  @Deprecated(
+    "AssetStorage requires type to find dependencies.",
+    replaceWith = ReplaceWith("AssetStorage.getDependencies")
+  )
   override fun getDependencies(fileName: String): GdxArray<String> = GdxArray.with<String>().also {
     logger.error("Not supported AssetManagerWrapper.getDependencies called by AssetLoader.")
   }
 
-  @Deprecated("AssetStorage requires type to find reference count.",
-    replaceWith = ReplaceWith("AssetStorage.getReferenceCount"))
+  @Deprecated(
+    "AssetStorage requires type to find reference count.",
+    replaceWith = ReplaceWith("AssetStorage.getReferenceCount")
+  )
   override fun getReferenceCount(fileName: String): Int = 0.also {
     logger.error("Not supported AssetManagerWrapper.getReferenceCount called by AssetLoader.")
   }