+++ async/src/test/kotlin/ktx/async/dispatchersTest.kt
@@ -5,10 +5,6 @@ import com.badlogic.gdx.utils.Timer
 import com.badlogic.gdx.utils.async.AsyncExecutor
 import com.nhaarman.mockitokotlin2.verify
 import io.kotlintest.mock.mock
-import java.util.concurrent.CompletableFuture
-import java.util.concurrent.atomic.AtomicBoolean
-import java.util.concurrent.atomic.AtomicInteger
-import java.util.concurrent.atomic.AtomicLong
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.Runnable
 import kotlinx.coroutines.async
@@ -18,7 +14,12 @@ import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNotSame
 import org.junit.Assert.assertSame
+import org.junit.Assert.assertTrue
 import org.junit.Test
+import java.util.concurrent.CompletableFuture
+import java.util.concurrent.atomic.AtomicBoolean
+import java.util.concurrent.atomic.AtomicInteger
+import java.util.concurrent.atomic.AtomicLong
 
 /**
  * Base class with coroutine dispatcher tests.
@@ -68,12 +69,13 @@ abstract class CoroutineDispatcherTest : AsyncTest() {
     }
 
     // Then:
-    assert(executionTime.get() in 50L..70L)
     assertSame(initialThread.get(), finalThread.get())
     assertNotSame(Thread.currentThread(), finalThread.get())
     if (!isConcurrent) {
       assertSame(executorThread, finalThread.get())
     }
+    // Tolerance adjusted for the testing environment:
+    assertTrue("${executionTime.get()} must be around 50 millis.", executionTime.get() in 45L..200L)
   }
 
   @Test
@@ -87,13 +89,18 @@ abstract class CoroutineDispatcherTest : AsyncTest() {
     val executionTime = AtomicLong()
 
     // When:
-    tested.invokeOnTimeout(50L, Runnable {
-      executionTime.set(System.currentTimeMillis() - start)
-    }, GlobalScope.coroutineContext)
+    tested.invokeOnTimeout(
+      50L,
+      Runnable {
+        executionTime.set(System.currentTimeMillis() - start)
+      },
+      GlobalScope.coroutineContext
+    )
 
     // Then:
     delay(100L)
-    assert(executionTime.get() in 50L..150L)
+    // Tolerance adjusted for the testing environment:
+    assertTrue("${executionTime.get()} must be around 50 millis.", executionTime.get() in 45L..200L)
   }
 
   @Test