+++ i18n/src/test/kotlin/ktx/i18n/I18nTest.kt
@@ -3,13 +3,13 @@ package ktx.i18n
 import com.badlogic.gdx.files.FileHandle
 import com.badlogic.gdx.utils.I18NBundle
 import io.kotlintest.mock.mock
-import java.io.File
 import ktx.i18n.I18nTest.BundleEnum.key
 import ktx.i18n.I18nTest.BundleEnum.keyWithArgument
 import org.junit.After
 import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Test
+import java.io.File
 
 /**
  * Tests internationalization-related classes and functions stored in *i18n.kt*.