+++ tiled/src/main/kotlin/ktx/tiled/mapLayers.kt
@@ -1,6 +1,7 @@
 package ktx.tiled
 
 import com.badlogic.gdx.maps.MapLayer
+import com.badlogic.gdx.maps.MapLayers
 import com.badlogic.gdx.maps.MapProperties
 
 /**
@@ -11,7 +12,7 @@ import com.badlogic.gdx.maps.MapProperties
  * @throws MissingPropertyException If the property is not defined.
  */
 inline fun <reified T> MapLayer.property(key: String): T = properties[key, T::class.java]
-    ?: throw MissingPropertyException("Property $key does not exist for layer $name")
+  ?: throw MissingPropertyException("Property $key does not exist for layer $name")
 
 /**
  * Extension method to directly access the [MapProperties] of a [MapLayer]. The type is automatically
@@ -38,3 +39,13 @@ inline fun <reified T> MapLayer.propertyOrNull(key: String): T? = properties[key
  * @return true if the property exists. Otherwise false.
  */
 fun MapLayer.containsProperty(key: String) = properties.containsKey(key)
+
+/**
+ * Returns **true** if and only if the [MapLayers] collection is empty.
+ */
+fun MapLayers.isEmpty() = this.count <= 0
+
+/**
+ * Returns **true** if and only if the [MapLayers] collection is not empty.
+ */
+fun MapLayers.isNotEmpty() = this.count > 0